Caleb: made changes per Rafaels requests
